Drop the implicit FIRST enum aliases and fix mixed enum operations
duplicate-enum reported 25 X_FIRST = 0 aliases that repeat an enum's first
real value: 22 in TD and three in RA's trigger and team enums. The project's
magic_enum rule already forbids such aliases, so they are gone and their
uses name the first enumerator.

The enum conversion diagnostics found comparisons against another enum's
equal-valued constant (Physics() against RESULT_NONE instead of IMPACT_NONE,
trigger actions against ACTION_NONE, serial packets against NET_FILE_CHUNK),
facing values added to animation IDs, and key numbers masked with the
WWKEY_* modifier bits. The modifier bits are flags rather than an
enumeration, so they became integer constants; the rest now name the right
constant or cast explicitly. Values are unchanged throughout.

The full strict build of both games is clean and all 242 tests pass.

+++ src/ra/techno.cc
@@ -3313,12 +3313,13 @@ BulletClass* TechnoClass::Fire_At(TARGET target, int which) {
     AnimType a = weapon->Anim;
     switch (a) {
       case ANIM_GUN_N:
-        a = static_cast<AnimType>(a + Dir_Facing(Fire_Direction()));
+        a = static_cast<AnimType>(
+            a + static_cast<int>(Dir_Facing(Fire_Direction())));
         break;
 
       case ANIM_SAM_N:
-        a = static_cast<AnimType>(ANIM_SAM_N +
-                                  Dir_Facing(PrimaryFacing.Current()));
+        a = static_cast<AnimType>(
+            ANIM_SAM_N + static_cast<int>(Dir_Facing(PrimaryFacing.Current())));
         break;
     }
